refactor: extract app state into app.rs
